fix: android styles
