+++ examples/stackflow-spa/src/stackflow/theme.ts
@@ -1,0 +1,3 @@
+export const theme = /iphone|ipad|ipod/i.test(window.navigator.userAgent.toLowerCase())
+  ? "cupertino"
+  : "android";
