+++ packages/recipe/lib/screen.mjs
@@ -16,10 +16,6 @@ const screenSlotNames = [
   [
     "edge",
     "screen__edge"
-  ],
-  [
-    "spacer",
-    "screen__spacer"
   ]
 ];
